+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
       and can be overridden when using `Probe`.
 - **FileProperties**: `FileProperties::new`
 - Debug logging via the [log](https://crates.io/crates/log) crate for exposing recoverable errors.
+- **Error**: `ErrorKind::SizeMismatch`
 
 ### Changed
 - **ID3v2**: Frame/tag flags with optional additional data are now `Option<T>` instead of `(bool, T)`